Updated SQLite connector, all queries pass, statements need testing.

+++ Source/Interface/Src/DatabaseLoggerConsole.cpp
@@ -456,7 +456,7 @@ BEGIN_NAMESPACE_DATABASE
 	{
 		if ( p_strToLog.find( STR( '\n' ) ) != String::npos )
 		{
-			StringArray array = CStrUtils::Split( p_strToLog, STR( "\n" ), std::count( p_strToLog.begin(), p_strToLog.end(), STR( '\n' ) ) + 1 );
+			StringArray array = CStrUtils::Split( p_strToLog, STR( "\n" ), uint32_t( std::count( p_strToLog.begin(), p_strToLog.end(), STR( '\n' ) ) + 1 ) );
 			auto && lastIt = array.begin() + array.size() - 1;
 			String last = *lastIt;
 			array.erase( lastIt );
